Move isAllowToImportPackage() function to util.ts
